+++ custom_components/greg/const.py
@@ -1,8 +1,8 @@
 """Constants for Greg integration."""
 
 DOMAIN = "greg"
-VERSION = "1.5.5"
-VERSION_DISPLAY = "v1.5.5"
+VERSION = "1.6.0b1"
+VERSION_DISPLAY = "v1.6.0b1"
 
 # Config keys
 CONF_VIBRATION_SENSOR = "vibration_sensor"